moving to v. 16 Mar 2018

+++ src/REPLICA/fix_event_prd.cpp
@@ -15,8 +15,8 @@
    Contributing author: Mike Brown (SNL)
 ------------------------------------------------------------------------- */
 
-#include "stdlib.h"
-#include "string.h"
+#include <stdlib.h>
+#include <string.h>
 #include "fix_event_prd.h"
 #include "atom.h"
 #include "update.h"
@@ -91,8 +91,8 @@ void FixEventPRD::restart(char *buf)
   double *list = (double *) buf;
 
   event_number = static_cast<int> (list[n++]);
-  event_timestep = static_cast<int> (list[n++]);
-  clock = static_cast<int> (list[n++]);
+  event_timestep = static_cast<bigint> (list[n++]);
+  clock = static_cast<bigint> (list[n++]);
   replica_number = static_cast<int> (list[n++]);
   correlated_event = static_cast<int> (list[n++]);
   ncoincident = static_cast<int> (list[n++]);
